Move rangeLength method to utils
